Update Tensorflow, gRPC, Protobuf dependencies (#868)
* Update Tensorflow to latest, finally update grpcio/protobuf


* Lint issue fix and missing tf reference


* pyzmq version fixed

* fix taskrunner tests for windows


* fix taskrunner test syntax for windows


* adding user option to workspace pip install requirements for windows


* fix windows CI test


* testing virtual env for windows github actions


* testing virtual env for windows github actions


* testing virtual env for windows github actions


* testing venv for windows


* test venv for windows

* test venv for windows

* Added new KerasSerializer. Fixed other Interactive API experiments

* Update taskrunner.yml

* Update taskrunner.yml

* Update workspace.py

* Update workspace.py

* Update taskrunner.yml

* Remove get_model import from global namespace so dependencies are not loaded into memory unnecessarily (breaking windows build)

* Refactoring and cleaning up imports to support Windows install

* Fixed logger import paths

* Fix missing imports

* Fix native import

* Fix lint errors

* Fix keras optimizer patch. Remove irrelevant unit test

* Format logs in UTF-8 for windows

* Update interactive-kvasir.yml

* Consolidate github actions python versions to single file

* Update python versions

* Update python versions

* Update python versions

* Reduce # of DataLoader workers for Pytorch Kvasir CI test

* Fix Windows encoding

* Fix Windows encoding and limit rounds so Github Actions CI doesn't run out of memory


* Fix windows encoding

* Fix Windows encoding

---------
