+++ .github/workflows/taskrunner.yml
@@ -12,19 +12,24 @@ permissions:
 
 jobs:
   build:
-
     strategy:
       matrix:
        os: ['ubuntu-latest', 'windows-latest']
+       python-version: ['3.8','3.9','3.10','3.11']
     runs-on: ${{ matrix.os }}
-
     steps:
     - uses: actions/checkout@v3
-    - name: Set up Python 3.8
-      uses: actions/setup-python@v3
+    - name: Set up Python 
+      uses: actions/setup-python@v4
       with:
-        python-version: "3.8"
-    - name: Install dependencies
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies ubuntu
+      if: matrix.os == 'ubuntu-latest'
+      run: |
+        python -m pip install --upgrade pip
+        pip install .
+    - name: Install dependencies windows
+      if: matrix.os == 'windows-latest'
       run: |
         python -m pip install --upgrade pip
         pip install .
