--- .github/workflows/taskrunner_python_3.9.yml
@@ -1,33 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: TaskRunner (Python 3.9)
-
-on:
-  pull_request:
-    branches: [ develop ]
-
-permissions:
-  contents: read
-
-jobs:
-  build:
-
-    strategy:
-      matrix:
-       os: ['ubuntu-latest', 'windows-latest']
-    runs-on: ${{ matrix.os }}
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python 3.9
-      uses: actions/setup-python@v3
-      with:
-        python-version: "3.9"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install .
-    - name: Test TaskRunner API
-      run: |
-        python -m tests.github.test_hello_federation --template keras_cnn_mnist --fed_workspace aggregator --col1 col1 --col2 col2 --rounds-to-train 3
