+++ openfl-tutorials/interactive_api/Tensorflow_Word_Prediction/workspace/requirements.txt
@@ -1,2 +1,2 @@
-tensorflow==2.11.1
+tensorflow==2.13
 numpy==1.22.2
