+++ openfl-workspace/keras_nlp/requirements.txt
@@ -1,1 +1,1 @@
-tensorflow==2.11.1
+tensorflow==2.13
