+++ openfl-workspace/keras_nlp_gramine_ready/requirements.txt
@@ -1,1 +1,1 @@
-tensorflow-cpu==2.11.1
+tensorflow-cpu==2.13
