+++ openfl-workspace/tf_cnn_histology/requirements.txt
@@ -1,3 +1,3 @@
 pillow
-tensorflow==2.11.1
+tensorflow==2.13
 tensorflow-datasets
