+++ openfl/experimental/utilities/runtime_utils.py
@@ -52,7 +52,7 @@ def filter_attributes(ctx, f, **kwargs):
     if "include" in kwargs and "exclude" in kwargs:
         raise RuntimeError("'include' and 'exclude' should not both be present")
     elif "include" in kwargs:
-        assert type(kwargs["include"]) == list
+        assert type(kwargs["include"]) is list
         for in_attr in kwargs["include"]:
             if in_attr not in cls_attrs:
                 raise RuntimeError(
@@ -62,7 +62,7 @@ def filter_attributes(ctx, f, **kwargs):
             if attr not in kwargs["include"]:
                 delattr(ctx, attr)
     elif "exclude" in kwargs:
-        assert type(kwargs["exclude"]) == list
+        assert type(kwargs["exclude"]) is list
         for in_attr in kwargs["exclude"]:
             if in_attr not in cls_attrs:
                 raise RuntimeError(
