+++ openfl/native/native.py
@@ -19,7 +19,7 @@ import openfl.interface.workspace as workspace
 from openfl.federated import Plan
 from openfl.protocols import utils
 from openfl.utilities import add_log_level
-from openfl.utilities import split_tensor_dict_for_holdouts
+from openfl.utilities.split import split_tensor_dict_for_holdouts
 
 logger = getLogger(__name__)
 
@@ -103,7 +103,7 @@ def update_plan(override_config, plan=None, resolve=True):
         else:
             # TODO: We probably need to validate the new key somehow
             logger.info(f'Did not find {key} in config. Make sure it should exist. Creating...')
-        if type(val) == list:
+        if type(val) is list:
             for idx, v in enumerate(val):
                 flat_plan_config[f'{key}.{idx}'] = v
         else:
