+++ tests/github/interactive_api_director/experiments/tensorflow_mnist/envoy/sd_requirements.txt
@@ -1,1 +1,1 @@
-tensorflow==2.11.1
+tensorflow==2.13
